refactor: 경매 기능에 관한 패키지 구조 수정
