--- src/main/java/com/kcs3/panda/domain/auction/bid/bidsample.java
@@ -1,5 +1,0 @@
-package com.kcs3.panda.domain.auction.bid;
-
-public class bidsample
-{
-}
