+++ src/main/java/com/kcs3/panda/domain/auction/controller/AuctionListController.java
@@ -1,0 +1,4 @@
+package com.kcs3.panda.domain.auction.controller;
+
+public class AuctionListController {
+}
