+++ src/main/java/com/kcs3/panda/domain/auction/dto/AuctionListDto.java
@@ -1,0 +1,4 @@
+package com.kcs3.panda.domain.auction.dto;
+
+public class AuctionListDto {
+}
