+++ src/main/java/com/kcs3/panda/domain/auction/entity/AuctionProgressItem.java
@@ -1,13 +1,9 @@
-package com.kcs3.panda.domain.auction.board.entity;
+package com.kcs3.panda.domain.auction.entity;
 
 import com.kcs3.panda.domain.model.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
 import jakarta.persistence.OneToOne;
 import java.time.LocalDateTime;
 import lombok.Data;
@@ -16,19 +12,23 @@ import lombok.EqualsAndHashCode;
 
 @EqualsAndHashCode(callSuper = true)
 @Entity
-public class AuctionCompleteItem extends BaseEntity {
+@Data
+public class AuctionProgressItem extends BaseEntity {
+
+
     @OneToOne(fetch = FetchType.LAZY, optional = true)
-    @JoinColumn(name = "itemId", nullable = true)
     private Item item;
     @Column(nullable = false)
     private String ItemTitle;
-    @Column(nullable=false)
+    @Column(nullable = false)
     private int starPrice;
     private int buyNowPrice;
     private LocalDateTime bidFinishTime;
-    @Column(nullable=false)
+    @Column(nullable = false)
     private String location;
-    @Column(nullable=false)
+    @Column(nullable = false)
     private int maxPrice;
     private String maxPersonID;
+
+
 }
