+++ src/main/java/com/kcs3/panda/domain/auction/entity/Item.java
@@ -1,4 +1,4 @@
-package com.kcs3.panda.domain.auction.board.entity;
+package com.kcs3.panda.domain.auction.entity;
 
 import com.kcs3.panda.domain.model.BaseEntity;
 import com.kcs3.panda.domain.user.entity.User;
@@ -6,7 +6,6 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
