--- src/main/java/com/kcs3/panda/domain/auction/list/controller/AuctionListController.java
@@ -1,4 +1,0 @@
-package com.kcs3.panda.domain.auction.list.controller;
-
-public class AuctionListController {
-}
