--- src/main/java/com/kcs3/panda/domain/auction/list/dto/AuctionListDto.java
@@ -1,4 +1,0 @@
-package com.kcs3.panda.domain.auction.list.dto;
-
-public class AuctionListDto {
-}
