--- src/main/java/com/kcs3/panda/domain/auction/list/entity/AuctionList.java
@@ -1,4 +1,0 @@
-package com.kcs3.panda.domain.auction.list.entity;
-
-public class AuctionList {
-}
