--- src/main/java/com/kcs3/panda/domain/auction/list/repository/AuctionListRepository.java
@@ -1,4 +1,0 @@
-package com.kcs3.panda.domain.auction.list.repository;
-
-public interface AuctionListRepository {
-}
