--- src/main/java/com/kcs3/panda/domain/auction/list/service/AuctionListService.java
@@ -1,4 +1,0 @@
-package com.kcs3.panda.domain.auction.list.service;
-
-public class AuctionListService {
-}
