+++ src/main/java/com/kcs3/panda/domain/auction/repository/AuctionListRepository.java
@@ -1,0 +1,4 @@
+package com.kcs3.panda.domain.auction.repository;
+
+public interface AuctionListRepository {
+}
