+++ src/main/java/com/kcs3/panda/domain/auction/service/AuctionListService.java
@@ -1,0 +1,4 @@
+package com.kcs3.panda.domain.auction.service;
+
+public class AuctionListService {
+}
